+++ placement/src/main/java/com/cdac/placement/service/FacultyService.java
@@ -5,10 +5,10 @@ import com.cdac.placement.model.Faculty;
 import java.util.List;
 
 public interface FacultyService {
-    public Faculty getFaculty(int facultyId);
+    public Faculty getFaculty(Long facultyId);
     public List<Faculty> getAllFaculties();
     public boolean deleteAllFaculties();
     public boolean addFaculty(Faculty faculty);
     public boolean updateFaculty(Faculty faculty);
-    public boolean deleteFaculty(Integer facultyId);
+    public boolean deleteFaculty(Long facultyId);
 }
