+++ placement/src/main/java/com/cdac/placement/service/MentorService.java
@@ -6,10 +6,10 @@ import java.util.List;
 
 public interface MentorService {
 
-    public Mentor getMentor(int mentorId);
+    public Mentor getMentor(Long mentorId);
     public List<Mentor> getAllMentors();
     public boolean deleteAllMentors();
     public boolean addMentor(Mentor mentor);
     public boolean updateMentor(Mentor mentor);
-    public boolean deleteMentor(Integer mentorId);
+    public boolean deleteMentor(Long mentorId);
 }
