+++ placement/src/main/java/com/cdac/placement/service/StudentService.java
@@ -6,11 +6,11 @@ import java.util.List;
 
 public interface StudentService {
 
-    public Student getStudent(int studentId);
+    public Student getStudent(Long studentId);
     public List<Student> getAllStudents();
     public boolean deleteAllStudents();
     public boolean addStudent(Student student);
     public boolean updateStudent(Student student);
-    public boolean deleteStudent(Integer studentId);
+    public boolean deleteStudent(Long studentId);
 
 }
